+++ CHANGELOG.md
@@ -23,8 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - A new `pop_analysis` kwarg to `quacc.schemas.cclib.summarize_run`.
 - New functions `quacc.util.dict.clean_dicts` and `quacc.util.dicts.sort_dict`.
 - Added support for Python 3.10.
-- Added a name VASP setting `MIN_VASP_VERSION`.
-- Added a new VASP swap for `EFERMI = "MIDGAP"` if `MIN_VASP_VERSION >= 6.4`.
+- Added a name VASP setting `VASP_MIN_VERSION`.
+- Added a new VASP swap for `EFERMI = "MIDGAP"` if `VASP_MIN_VERSION >= 6.4`.
 
 ### Changed